requirements.txt
Django==5.2.4
djangorestframework==3.16.1
django-filter==25.2
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.2.1
openai==2.21.0
django-cors-headers==4.9.0